requirements.txt
PyOpenGL
PyOpenGL_accelerate
opencv-python
glfw
numpy
